+++ datanode/gateway/graphql/helpers.go
@@ -30,6 +30,15 @@ func safeStringUint64(input string) (uint64, error) {
 	return i, nil
 }
 
+func safeStringInt64(input string) (int64, error) {
+	i, err := strconv.ParseInt(input, 10, 64)
+	if err != nil {
+		// A conversion error occurred, return the error
+		return 0, fmt.Errorf("invalid input string for int64 conversion %s", input)
+	}
+	return i, nil
+}
+
 func secondsTSToDatetime(timestampInSeconds int64) string {
 	return vegatime.Format(vegatime.Unix(timestampInSeconds, 0))
 }